# docs/swagger.yaml
definitions:
  project.DataPacket:
    properties:
      data:
        items:
          type: integer
        type: array
      direction:
        type: string
      displayData:
        type: string
      endOffset:
        type: integer
      guid:
        type: string
      id:
        type: integer
      modified:
        type: boolean
      requestID:
        type: integer
      startOffset:
        type: integer
      time:
        type: integer
    type: object
...
  project.InjectOperationRequestPart:
    properties:
      inject:
        type: boolean
      requestPart:
        type: string
    type: object
...
  project.PartialRequestResponseData:
    properties:
      data:
        type: string
      from:
        type: integer
      to:
        type: integer
    type: object
  project.Request:
    properties:
      error:
        type: string
      guid:
        type: string
      hash:
        type: string
      notes:
        type: string
      objectType:
        type: string
      payloads:
        type: string
      protocol:
        type: string
      requestSize:
        type: integer
      responseContentType:
        type: string
      responseSize:
        type: integer
      responseStatusCode:
        type: integer
      responseTime:
        type: integer
      scanID:
        type: string
      time:
        type: integer
      url:
        type: string
      verb:
        type: string
    type: object
  project.RequestDifference:
    properties:
      request:
        description: 1 for request number one, 2 for request number two, 0 for both
        type: integer
      text:
        type: string
    type: object
  project.RequestResponseContents:
    properties:
      combinedContentLength:
        type: integer
      dataPackets:
        items:
          $ref: '#/definitions/project.DataPacket'
        type: array
      largeResponse:
        type: boolean
      mimeType:
        type: string
      modified:
        type: boolean
      modifiedRequest:
        type: string
      modifiedResponse:
        type: string
      protocol:
        type: string
      request:
        type: string
      response:
        type: string
      url:
        type: string
    type: object
  project.RequestSummary:
    properties:
      guid:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      hostname:
        type: string
      protocol:
        type: string
      requestData:
        type: string
      siteMapPath:
        type: string
      splitRequest:
        items:
          $ref: '#/definitions/project.InjectOperationRequestPart'
        type: array
      url:
        type: string
    type: object
...
host: localhost
info:
  contact:
    email: support@forensant.com
    name: API Support
    url: https://forensant.com/support
  description: This provides the common functions which are relied upon by the Proximity
    frontends.
  license:
    name: Commercial
    url: https://proximity.forensant.com/terms
  termsOfService: https://forensant.com/terms/
  title: Proximity Core
  version: "1.0"
paths:
...
  /requests:
    get:
      description: gets a list of all requests
      parameters:
      - description: Scan ID, can be multiple separated by semi-colons
        in: query
        name: scanid
        type: string
      - description: Only show requests which contain the filter string in the url,
          request, response, etc
        in: query
        name: filter
        type: string
      - description: Only show requests which contain the given string in the URL
        in: query
        name: url_filter
        type: string
      - description: Column to sort by (default time)
        in: query
        name: sort_col
        type: string
      - description: Column direction to sort by (default asc)
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.Request'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Requests
      tags:
      - Requests
  /requests/{base_guid}/compare/{compare_guid}:
    get:
      description: compares two requests and returns the differences
      parameters:
      - description: Base Request guid
        in: path
        name: base_guid
        required: true
        type: string
      - description: Request to Compare guid
        in: path
        name: compare_guid
        required: true
        type: string
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.RequestDifference'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Compare Two Requests
      tags:
      - Requests
  /requests/{guid}:
    get:
      description: gets a specific request
      parameters:
      - description: The GUID of the request to fetch
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.RequestSummary'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get A Request
      tags:
      - Requests
  /requests/{guid}/contents:
    get:
      description: gets the full request and response of a given request
      parameters:
      - description: Request GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.RequestResponseContents'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Request and Response
      tags:
      - Requests
  /requests/{guid}/notes:
    patch:
      description: updates a specific request's notes
      parameters:
      - description: The GUID of the request to update
        in: path
        name: guid
        required: true
        type: string
      - description: The notes for the request
        in: body
        name: notes
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Request Notes
      tags:
      - Requests
  /requests/{guid}/partial_data:
    get:
      description: gets part of the request/response. will attempt to return at least
        5MB of data to cache
      parameters:
      - description: Request guid
        in: path
        name: guid
        required: true
        type: string
      - description: Offset to request from
        in: query
        name: from
        required: true
        type: integer
      produces:
      - text/text
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.PartialRequestResponseData'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Request/Response Data
      tags:
      - Requests
  /requests/{guid}/payloads:
    patch:
      description: sets the payloads associated with a specific request
      parameters:
      - description: The GUID of the request to update
        in: path
        name: guid
        required: true
        type: string
      - description: A JSON Object containing the payloads in {'key':'value'} format
        in: body
        name: payloads
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Request Payloads
      tags:
      - Requests
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"